pyproject.toml
[project]
name = "chess-reporter"
version = "0.2.1"
description = "Processes PGN files and analyzes each position using Stockfish. Multiple engine instances run in parallel (default: 5, always odd) and results are aggregated for robustness. Data is stored in DuckDB and explored via JupyterLab."
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "duckdb>=1.4.4",
    "jupyterlab>=4.5.5",
    "loguru>=0.7.3",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "pydantic>=2.12.5",
    "python-chess>=1.999",
    "sqlglot>=29.0.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pyright>=1.1.0",
    "pytest>=9.0.2",
    "ruff>=0.15.5",
]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
exclude = ["notebooks/"]

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pyright]
typeCheckingMode = "standard"
pythonVersion = "3.13"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/chess_reporter"]

[project.scripts]
clean = "chess_reporter._scripts.clean:main"
